Commit - Print Changes - Please Read Description
The Print class now only takes a MainScreen as its parameter. For the
Print method, please pass the graphics of the object to be printed as
the first parameter. For the Edit Action item Screen, this would be the
content pane of the window. Everything else has remained the same.
Future changes are possible and probable.